backend/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.29.0
openai==1.74.0
pydantic==2.6.4
beautifulsoup4==4.12.3
python-dotenv==1.0.1
